+++ ReleaseNotes.md
@@ -22,3 +22,4 @@
 - Add MileFreeMemory function.
 - Add MileGetTickCount function.
 - Add MileCreateThread function.
+- Add MileGetNumberOfHardwareThreads function.